% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{add_network_type}
\alias{add_network_type}
\title{Annotate Flowpaths and Divides with Type Info}
\usage{
add_network_type(network_list)
}
\arguments{
\item{network_list}{List of \code{flowpaths} and \code{divides} (sf objects).}
}
\value{
Updated \code{network_list} with type info columns.
}
\description{
Adds boolean flags \code{has_divide} to flowpaths and \code{has_flowpath} to divides.
}
