% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_distribution.R
\name{assign_id}
\alias{assign_id}
\title{Assign Group IDs Based on Ideal Size}
\usage{
assign_id(x, athres)
}
\arguments{
\item{x}{Numeric vector.}

\item{athres}{Numeric. Threshold cumulative sum before starting new group.}
}
\value{
Integer vector of group IDs.
}
\description{
Groups sequential values until their sum exceeds the threshold.
}
