% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_distribution.R
\name{pinch_sides}
\alias{pinch_sides}
\title{Merge Small Groups on the Edges}
\usage{
pinch_sides(x, ind, thres)
}
\arguments{
\item{x}{Vector of values.}

\item{ind}{Grouping index for x.}

\item{thres}{Numeric threshold for merging.}
}
\value{
Updated group index.
}
\description{
Merges the first and last group if they're smaller than the threshold.
}
